+++ src/common/client.ts
@@ -798,6 +798,8 @@ export class Client extends EventEmitter {
                 alarm.prevTime = now;
                 //save as if temp alarm as execute trigger advances state and temp alarms will need different state shifts
                 const state = parent.state;
+                this._input.lastTriggered = alarm.pattern;
+                //_Triggered = string.Format("{0}:{1}:{2}", Fired.Hour, Fired.Minute, Fired.Second);
                 this._input.ExecuteTrigger(trigger, [alarm.pattern], false, -a, null, null, parent);
                 if (state !== parent.state)
                     alarm.restart = Date.now();